Multiple New Features and contiue to code and refine frame work. Completely re-coded the setup as some where finding it a little confusing. Set up no longer requires you to touch a file. All done in game step by step. Added goes form game logic to Lobby Manager, GameManager, Game. New Listeners to help separate SCB for effecting other plugins. Started a custom messaging system to separate all chat even on a mixed mode server. Now as a hard dependancy of Vault which is used for Chat, Perms and Eco when the time comes. Days of research and testing The optimum way of multithreading various events and tasks, was well worth it :) Hal way though triple jump. Join Lobby and Join Game signs done. Inventory management well one the way. Plus more in this commit.

+++ src/main/java/com/relicum/scb/objects/signs/interfaces/ISignLocation.java
@@ -8,10 +8,10 @@ package com.relicum.scb.objects.signs.interfaces;
  */
 public interface ISignLocation {
 
-    int getX();
+    double getX();
 
-    int getY();
+    double getY();
 
-    int getZ();
+    double getZ();
 
 }
